[473114] Working around an issue related to finding views
In Mars the the findView method does not always finds a selected view
(see https://bugs.eclipse.org/bugs/show_bug.cgi?id=468948), that caused
the Query Explorer to throw an NPE when the Query Explorer is opened. As
a workaround, IncQuery tries first the currently selected view in
addition to searching by ID that seems to fix the issue.